+++ src/components/Box/index.tsx
@@ -9,6 +9,7 @@ export const Box = ({
   direction = "column",
   justify = "baseline",
   align = "stretch",
+  wrap,
   ...props
 }: (
   | { gap: "xs" | "sm" | "md" | "lg" | "xl" | "xxl"; customValue?: undefined }
@@ -19,11 +20,13 @@ export const Box = ({
   direction?: "column" | "row" | "column-reverse" | "row-reverse";
   justify?: "center" | "space-between" | "space-around" | "end" | "left" | "right" | "baseline";
   align?: "center" | "end" | "start" | "baseline" | "stretch";
+  wrap?: "nowrap" | "wrap";
 }) => {
   const customStyle = {
     "--Box-direction": direction,
     "--Box-justify": justify,
     "--Box-align": align,
+    ...(wrap ? { flexWrap: wrap } : {}),
   } as React.CSSProperties;
 
   return (