+++ src/constants/settings.ts
@@ -37,6 +37,7 @@ export enum Routes {
   PAYMENTS = "/payments",
   DISTRIBUTION_ACCOUNT = "/distribution-account",
   WALLET_PROVIDERS = "/wallet-providers",
+  WALLET_PROVIDERS_NEW = "/wallet-providers/new",
   WALLET = "/wallet",
   WALLET_HOME = "/wallet/home",
   ANALYTICS = "/analytics",