+++ src/helpers/refreshWalletSessionToken.ts
@@ -1,5 +1,6 @@
-import { AppDispatch } from "@/store";
 import { refreshWalletTokenAction } from "@/store/ducks/walletAccount";
+
+import { AppDispatch } from "@/store";
 
 export const refreshWalletSessionToken = (dispatch: AppDispatch) => {
   dispatch(refreshWalletTokenAction());